// model/apiClient.ts
/**
 * MCC Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ApiClient {
    'createdAt': Date;
    /**
    * description about the api client
    */
    'description'?: string;
    /**
    * the client id
    */
    'clientId': string;
    /**
    * the client id
    */
    'clientSecret': string;
    'lastUsedAt'?: Date;
    /**
    * Scope
    */
    'scope': Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "clientId",
            "baseName": "client_id",
            "type": "string"
        },
        {
            "name": "clientSecret",
            "baseName": "client_secret",
            "type": "string"
        },
        {
            "name": "lastUsedAt",
            "baseName": "last_used_at",
            "type": "Date"
        },
        {
            "name": "scope",
            "baseName": "scope",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return ApiClient.attributeTypeMap;
    }
}