// model/apiClientDraft.ts
/**
 * MCC Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ApiClientDraft {
    /**
    * description about the api client
    */
    'description'?: string;
    /**
    * Scope
    */
    'scope'?: Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "scope",
            "baseName": "scope",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return ApiClientDraft.attributeTypeMap;
    }
}