// model/commitData.ts
/**
 * MCC Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CommitDataAuthor } from './commitDataAuthor';

export class CommitData {
    'commit': string;
    'parents'?: Array<string>;
    'subject': string;
    'author': CommitDataAuthor;
    'committer': CommitDataAuthor;
    'createdAt'?: Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "commit",
            "baseName": "commit",
            "type": "string"
        },
        {
            "name": "parents",
            "baseName": "parents",
            "type": "Array<string>"
        },
        {
            "name": "subject",
            "baseName": "subject",
            "type": "string"
        },
        {
            "name": "author",
            "baseName": "author",
            "type": "CommitDataAuthor"
        },
        {
            "name": "committer",
            "baseName": "committer",
            "type": "CommitDataAuthor"
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return CommitData.attributeTypeMap;
    }
}