// model/commitDataPaginatorAllOf.ts
/**
 * MCC Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CommitData } from './commitData';

export class CommitDataPaginatorAllOf {
    'results'?: Array<CommitData>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "results",
            "baseName": "results",
            "type": "Array<CommitData>"
        }    ];

    static getAttributeTypeMap() {
        return CommitDataPaginatorAllOf.attributeTypeMap;
    }
}